- int to long in estimated number of results. - added try/catch block around initial fetch.
git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@1563 7ff1d41c-760d-0410-a7ff-a3a56f310b35

+++ carrot2/components/carrot2-local-core/src/org/carrot2/core/fetcher/SearchResult.java
@@ -22,11 +22,11 @@ import org.carrot2.core.clustering.RawDocument;
 public final class SearchResult
 {
     public final int at;
-    public final int totalEstimated;
+    public final long totalEstimated;
     public final Throwable error;
     public final RawDocument [] results;
 
-    public SearchResult(RawDocument [] results, int at, int totalEstimated)
+    public SearchResult(RawDocument [] results, int at, long totalEstimated)
     {
         this.results = results;
         this.totalEstimated = totalEstimated;
